Optimize frequently called global functions
When a function is called in a namespace,
the PHP interpreter has to check for it both in the current namespace
and the global namespace.
Opcache can't optimize it, because there's a chance an eval(),
require(), etc. may add the function to the current namespace,
shadowing the global namespace.

```
php -d xdebug.profiler_enable=1 \
  -d xdebug.profiler_output_dir=$PWD/tmp bin/psalm
```

Impact: Sped up from 3.33 to 3.30 (1% speedup)